Separated User/Social from other Tables with new app 'Details'
