+++ Accounts/admin.py
@@ -1,0 +1,6 @@
+from django.contrib import admin
+from .models import *
+# Register your models here.
+admin.site.register(User)
+admin.site.register(UserSocial)
+
