+++ Accounts/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class SmaConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'SMA'
+    name = 'Accounts'
