+++ Accounts/views.py
@@ -3,4 +3,4 @@ from django.http import HttpResponse
 # Create your views here.
 
 def index(request):
-    return HttpResponse ("Welcome, this is the SMA app")
+    return HttpResponse("Welcome, this is the SMA app")
