+++ sma_project/settings.py
@@ -13,6 +13,10 @@ https://docs.djangoproject.com/en/3.2/ref/settings/
 from pathlib import Path
 import django_heroku
 
+# Extending User Model Using a Custom Model Extending AbstractUser
+# We have to update our settings.py defining the AUTH_USER_MODEL property.
+AUTH_USER_MODEL = 'Accounts.User'
+
 # Build paths inside the project like this: BASE_DIR / 'subdir'.
 BASE_DIR = Path(__file__).resolve().parent.parent
 
@@ -40,7 +44,8 @@ INSTALLED_APPS = [
     'django.contrib.staticfiles',
 
     #application created by Kelvin
-    'SMA',
+    'Accounts',
+    'Details',
 
 ]
