+++ sma_project/urls.py
@@ -19,5 +19,6 @@ from django.urls import path, include
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('', include('SMA.urls'))
+    path('', include('Accounts.urls')),
+
 ]
